Rebuilt implant categorisation, to support reused links.
All implants stagers now create a 'generated' implant with a unique key based
on their stager template. This has affected most of the components in within
Fudge, and requires further testing to ensure any bugs, and/or race conditions
are identified.
Code refactor is now required, and further checks to ensure that user access
control is adheared to. This may require redesign as well