Added AJAX checkin feature, added graph page with placeholder & password reset complete
Added the password reset on first login, and checkes on subsequent logins.

Numerous changes to the database file. This should be refactored to contain only database queries,
and a second file should be created to manipulate data for specific pages, allowing Database to be
cleaner, and the webapp endpoint to be more specific.

All endpoints need to be made consistent, and ensure that changes to endpoints do not cause bugs.