Implant Updated
The implant and inbuilt command PoC is now live.
The built in string '''::''' is a pre-fix for all builtins which the C2
server will validate, before registering commands.

JinjaFudge has also been updated to reflect this change, and will return the
result of the first builtin ''':: sys_info''' with <username>::<hostname>.

Small bugs removed and printing removed.