Adding initial files, and changes to listener functionality
As part of the C2 framework the admin must be able to perform the following
actions:

 - Configure and start new listener
 - Query all active listeners
 - Stop a given listener

This will be done through the 'ListenerManagement' class this is currently
passed through to the server application as a Flask config object.

A global page will be added to the top bar which shows active listeners.
the http listener will be active on port 5000 by default, however, this
will likely be removed in future.

Further updates to the roadmap will be required to facilite these changes.