Major Updates: HTTPS listener with user-supplied certificates partially implemented.
Added in HTTPS certificates for listeners. Stored in /Storage/.
Requires error checking to ensure they are present. HTTPS implants
 are not yet ready due to connections not closing as expected.

Added in functionality to save a local copy of the generated implant
to the database. Allows RT'ers to present the text version of the
implant to defenders for analysis. This is stored in the SQL DB
under the 'generated_implants' table.

Implant responses are now formatted correctly with newlines, making commands
like ipconfig easier to read.

Beacon time is now implemented and working correctly.

Cleaned up Dockerfile