Adding dev work for module loading into FudgeC2
Run using:
    :: load_module [module filename]

All modules should be stored within Storage/implant_resources/modules/

Currently only .ps1 files are checked, if not matching module is found
the LM payload will not generate.

Outstanding work needs to be done to validate if the initial command
submission is selecting a valid payload at the time of request.