Updates to how FC2 handles direct command execution. Previously commands such as below would fail, they now function as expected:     Get-Command -Module * | where {$_.Name -Like "*SmbC*"}
3 builtin commands have now been added/updated
 - :: load_module
 - :: exec_module [function]
 - :: list_modules

Loading modules improved - when a module is loaded it will return a list of the functions available.
To execute the functions of the loaded module Invoke-Expression is called  via the :: exec_module builtin command.
The user needs to pass a loaded function into this command.
If a user is not sure which modules have been loaded by the implant you can run the :: list_modules command to query loaded modules.

Webapp:
 - Minor tweaks throughout
 - Font change
 - Initial API endpoints tested